requirements.txt
requests~=2.32.3
python-dotenv~=1.0.1
black==23.9.1
isort==6.0.0